Define Parametric Mixture or Non-Mixture Cure Distribution

Define parametric mixture or not mixture cure distribution with given parameter values.

Usage

define_surv_cure(distribution, theta, ..., mixture = TRUE)

define_survival_cure(distribution, theta, ..., mixture = TRUE)

Arguments

distribution

A parametric survival distribution. See details for a listing of valid distributions.

theta

A numeric value representing cure fraction

...

Additional distribution parameters (see details section below)

mixture

a logical determining whether a mixture or non-mixture model is being defined.

Value

a surv_dist_cure object.

Details

Supported distributions are listed in the table below.

Distribution	Description	Parameters	Notes

"exp"	Exponential	rate

"lnorm"	Lognormal	meanlog, sdlog

"llogis"	Log-Logistic	shape, scale

"weibull"	Weibull (AFT)	shape, scale

"weibullPH"	Weibull (PH)	shape, scale

"gompertz"	Gompertz	shape, rate

"gamma"	Gamma	shape, scale

"gengamma"	Generalized Gamma (stable)	mu, sigma, Q	Parameterization from Prentice (1974)

"gengamma.orig"	Generalized Gamma (original)	shape, scale, k	Original parameterization from Stacy (1962)

"genf"	Generalized F (stable)	mu, sigma, Q, P	Stable reparameterization from Prentice (1975)

"genf.org"	Generalized F (original)	mu, sigma, s1, s2	Origninal parameterization described by Prentice (1975)

References

Stacy, E. W. (1962). A generalization of the gamma distribution. Annals of Mathematical Statistics 33:1187-92.

Prentice, R. L. (1974). A log gamma model and its maximum likelihood estimation. Biometrika 61(3):539-544.

R. L. Prentice (1975). Discrimination among some parametric models. Biometrika 62(3):607-614.

Examples

define_surv_cure(distribution = "exp", theta = 0.34, rate = .5)
#> An exponential mixture cure distribution (theta = 0.34, rate = 0.50).
define_surv_cure(
 distribution = "weibull",
 theta = 0.5, shape = 1.5,
 scale = 34.43,
 mixture = TRUE
)
#> A Weibull (AFT) mixture cure distribution (theta = 0.5, shape = 1.5, scale = 34.4).
# Deprecated alias included for backwards compatability with heRomod
define_survival_cure(distribution = "exp", theta = 0.24, rate = 0.023)
#> An exponential mixture cure distribution (theta = 0.240, rate = 0.023).